pyproject.toml
[tool.poetry]
name = "web-market"
version = "0.1.0"
description = ""
authors = ["Rotundyer <Gatzknow@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
envparse = "^0.2.0"
uvicorn = "^0.29.0"
asyncpg = "^0.29.0"
alembic = "=1.13.1"
psycopg2-binary = "2.9.9"
fastapi = "^0.111.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
python-dotenv = "^1.0.1"
requests = "^2.31.0"
sqlalchemy = "2.0.29"
sqlmodel = "^0.0.18"
aiohttp = "^3.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
